requirements.txt
Pillow>=9.0.0
tqdm>=4.0.0
watchdog>=2.1.0
